+++ priv/repo/migrations/20240319131049_drop_usage_tracking_configurations.exs
@@ -1,0 +1,7 @@
+defmodule Lightning.Repo.Migrations.DropUsageTrackingConfigurations do
+  use Ecto.Migration
+
+  def change do
+    drop table(:usage_tracking_configurations)
+  end
+end